Store new master key when calling srtp_update
The master key now needs to be stored in srtp srtp_stream_ctx_t.
The update_srtp function was not doing this so master key was
effective all zeros after srpt_update was called.
The code that manages the creation of new streams in side of
srtp_update is basically a copy of what is done in srtp_add_stream,
this code should be re factored to be the same to avoid such
inconsistency errors.