Restore AES-192 under BoringSSL
PR #93 added the SRTP_NO_AES192 macro and removed all references to AES-192
if it was set. The goal was to get libsrtp to compile with BoringSSL, which
didn't support AES-192 at the time (Feb 2015).

In the time since, BoringSSL has restored support for AES-192:
https://boringssl.googlesource.com/boringssl/+/1049e26f
https://boringssl.googlesource.com/boringssl/+/5dca031c

Remove the SRTP_NO_AES192 macro, since it's no longer needed.
Effectively reverts #93.