configure.in: Changes empty CFLAGS check to be (more) shell neutral.
The test for empty CFLAGS used the bash (and compatible) -z test to check
for a zero length string.

This changes it to the more shell neutral "x$foo" = "x" form that's also
used elsewhere in this file.

The AS_IF macro is the preferred way of expressing these tests.
Since there are a few of them in this file, such a rewrite would touch
large parts of it. Perhaps something to look into later.