configure.in/Makefile.in: Always generate and install pkg-config file.
The libsrtp2.pc file should always be generated and installed regardless
of whether pkg-config is installed on the build machine.

For instance, if a machine is dedicated to building packages, but doesn't
have pkg-config installed, that doesn't mean users of the finished package
doesn't want the pkg-config file.

The libsrtp2.pc file, generated from the libsrtp2.pc.in template, is only
dependent on the correct AC_SUBST() calls in configure.in. I.e. pkg-config
does not have to be installed to generate it.

This removes the entire HAVE_PKG_CONFIG concept and always generates and
installs the libsrtp2.pc file.