test/rtpw: Adds local words.txt and uses it.
The rtpw test program tried to use either a "words.txt" file (on Windows)
or "/usr/share/dict/words" on other platforms.

The "words.txt" file is not in the repository and "/usr/share/dict/words"
may not exist (may not have been installed).

To make the test program independent of files that may or may exist on a
given system, this adds a local "words.txt" file changes rtpw.c to always
use that file.

The collection of words in that file (250 for now) is generated from some
random random word generator (pun intended).