test/rtpw_test*.sh: Account for Windows executable and silence output.
This adds a ".exe" postfix to the rtpw executable used in the test
scripts when running under Cygwin.

It also pipes errors from the wait command to /dev/null to silence
ts' output. This would otherwise print a terminated message to stderr.

In the case of the "*_gcm.sh" script, the wait command was added, with
the stderr redirection. Without it, it would print the same terminated
message to stderr.