Report ssrc instead of srtp_stream_t in srtp_event_data_t
srtp_stream_t is an opaque type and unusable from the public api,
this made the entire event framework almost useless.
This change swaps the stream point with an ssrc, which is used as a key
in the public api. A uintptr_t is used instead of a uint32_t to ensure
the size of the srtp_event_data_t struct remains unchanged.

This change could potentially break existing code but the assumption
is that the number of uses of this api that are doing anything with this
value after the version 2.0 release approaches zero.