Simplify logging of errors and debug
Previously logging of debug and errors did not always work as expected.
Now there are three configure options to control output. The idea is not to
be overly complicated but to give some simple options that work.

--enable-log-stdout
This will out put all debug and error logging to stdout

--with-log-file=<filename>
This will output all debug and error logging to the given filename

--enable-debug-logging
This enable debug logging for all internal debug modules. This is very
verbose and should only be used for debugging and development.