configure.in: Lowers pkg-config version required.
The use of PKG_CHECK_MODULES_STATIC was introduced in 333fa84e8e alongside
requiring pkg-config 0.29 or later.

PKG_CHECK_MODULES_STATIC was introduced in May 2013 in respons to
pkg-config bug 19541:
<https://bugs.freedesktop.org/show_bug.cgi?id=19541>

Gathered from version history, it appears PKG_CHECK_MODULES_STATIC was
released with pkg-config 0.29 in September 2015.

The use of PKG_CHECK_MODULES_STATIC and the pkg-verison requirement aren't
really connected tough. 0.29 or later is only required for regenerating
the configure script, not really for using it.

And it has also become apparent that 0.29 is too new. Developers that
wants to make changes and regenerate the configure script might not have
it and therefore error out with a warning about PKG_CHECK_MODULES_STATIC
being undefined when running autoreconf.

This removes the pkg-config version requirement again.

And instead of using PKG_CHECK_MODULES_STATIC, it sets "$PKG_CONFIG" to
"$PKG_CONFIG --static" and uses the regular PKG_CHECK_MODULES macro.
This is in essence what PKG_CHECK_MODULES_STATIC does internally, and was
the recommended way of doing this prior to its introduction in 0.29.