Apply some formatting nits
Core of this is to limit the line length to at least 90 width where it
makes sense contextually.

Some whitespace fixups were applied as well, mainly replacing tabs with
spaces since this seems like the most consistently used white space.
This was done around points where this PR is touching the repo in order
to improve the readability slightly.