Issue #256 Seq number incorrectly masked for IV.
During the IV calculation, the sequence number has it's endianess
reversed before clearing the most significant bit, bit 31.
On little endian platforms, this results in the incorrect clearing
of bit 7. The fix is to clear the most significant bit prior to
swapping the byte order.
Please note that this change results in incompatibility with libsrtp
releases without this change.