Remove generated documentation
Committing the generated documentation caused issues when doing a
make clean, it would delete the documentation.
A new branch was created master_generated_doc that will be used to have
generated docs available in github, this can be updated with each new
release, reducing the need to have more commits in master.