test_srtp.c: Regression test for issue #256
This is simply a regression test and a model for how we can test the
generation of IVs without firing up a full crypto session.

This test should be agnostic to the endianness of the system since it
only deals with setting of single bytes for the postcondition IVs.

We check that each IV is generated as expected and we check that each IV
is distinct.