configure.in: Adds support for EXTRA_CFLAGS to be appended to CFLAGS.
If the user sets CFLAGS on the command line, configure will not set its
own set of default flags. Which makes sense. It's annoying when configure
scripts thinks they know better and overrides a user's optimization or
warning flags.

However, this also makes it impossible to just add something to CFLAGS
short of manually inspecting and copying the defaults on the command line
along with whatever flag(s) the user wanted to add.

This fixes it by adding support for EXTRA_CFLAGS that will be appended
to CFLAGS after all other checks are done.

This also reorders some of the program checks to get the "influential
environment variables" section of `configure --help` to look right and
removes a rouge "dnl" inside a comment block left by accident by a
earlier reformatting change