Detect and handle broken "OPENSSL_cleanse".
On some platforms (experienced this on arm64 with OpenSSL 1.0.2g), the
function `OPENSSL_cleanse` is broken and doesn't clear the memory, but
fills it with data. This breaks `octet_string_set_to_zero`.

This patch checks if `OPENSSL_cleanse` is broken during configure and if
it is, will use `srtp_cleanse` instead. Also a testcase is added to
`datatypes_driver.c` to check if `octet_string_set_to_zero` is working
correctly.