srtp.c: Fixes unaligned trailer access
All comments states that the trailer pointer should be 4-byte aligned
because 'packets' are always multiples of 32-bits.

But the address calculation for the trailer pointer includes tag length,
and tag length may not be a multiple of 4. See the
'srtp_crypto_policy_set*()' functions, where some have a tag length of 10
for instance.

Given the following trailer calculation:

trailer_p = (uint32_t *)((char *)hdr + *pkt_octet_len -
                        (tag_len + mki_size + sizeof(srtcp_trailer_t)));

and the following assumptions:
- hdr is correctly aligned
- mki_size = 0
- sizeof(srtcp_trailer_t) = 4
- tag_len = 10

Then the only variable in the offset calculation is pkg_octet_len, from
which we subtract 14 (which is not a multiple of 4).

Given pkg_octet_len = 46, the offset will be 32, which is aligned.
Given pkg_octet_len = 32, the offset will be 18, which is unaligned.

So even for tag lengths that aren't a multiple of 4, it will end up being
aligned some times, but not all the time.

What effect unaligned memory access has is architecture depended. Crashes,
huge performance penalties in the form of traps and fix-ups, or a almost
negligible CPU cache line miss. From a programming perspective, it's
undefined behaviour.

This fixes it by changing trailer access to use a intermediate variable
via 'memcpy()' instead of using the pointer directly.