support system NSS packages
Add support for useing pkgconfig to find system installed
versions of NSS, this is based on the current openssl
support.
This removes the need to specify --with-nss-dir but that
option is still available if a custom NSS build is required.

On osx brew installs nss to a seperate location to avoid possible
conflicts with firefox so --with-nss-dir or PKG_CONFIG_PATH needs
to be used.

Update travis scripts to use system packages.