switch NSS_NoDB_Init() to NSS_InitContext()
libsrtp uses nss in the context of a system
library. It is up to the application to call
NSS_XXX_Init(). Changing to NSS_InitContext()
allows libsrtp to initialize as much of nss as
it requires without effecting the application. It
also allows nss cleanup through the use of
NSS_ShutdownContext()

This fixes potential memory leaks.