support decoding base64 AEAD_AES_GCM key
The AEAD_AES_GCM key and salt is encoded in a base64 string
that includes padding. Checking the padding result does not
give any meaning in the normal case either, len & expected_len
are always checked to verify that the key is correct.

The full length of the input_key should be passed to
base64_string_to_octet_string so it can correctly deal with strings
that are padded.