configure.ac: Drop pkg-config usage when --with-openssl-dir is specified
When installing OpenSSL (using 'make install'), it's possible to specify a
INSTALL_PREFIX (this can also be set during configuration). This is
similar to the customary DESTDIR mechanism, but not quite.

DESTDIR will not duplicate the entire --prefix hierarchy, but OpenSSL's
INSTALL_PREFIX will..

Given a package "foo" configured with 'Configure --prefix=/opt/local':
- The customary 'make DESTDIR=/tmp/foo install' would install it to
  "tmp/foo/bin", "tmp/foo/lib", etc.
- OpenSSL's 'make INSTALL_PREFIX=/tmp/foo install' would install it to
  "tmp/foo/opt/local/bin", "tmp/foo/opt/local/lib" etc.

Package config files are generated based on the provided --prefix. So even
if installed to somewhere else, the package config file will still point
to the final destination provided during configuration.

This configure script would call pkg-config, if available, also
when --with-openssl-dir was used.

There are old and pedantic reasons for that, primarily relating to finding
the correct dependencies and avoiding specifying the library multiple
times (event though that doesn't hurt).

However, if the mentioned INSTALL_PREFIX mechanism was used to install
OpenSSL, pkg-config would add flags pointing to the --prefix location used
when building OpenSSL, not the provided --with-openssl-dir.

Since the crypto library is so ubiquitous, this would most likely end up
using the system library instead, if the rest of the configure checks
found the functions it's looking for in the system library.

If there happened to be a crypto library already at the given --prefix,
that would be used (again provided the configure checks passed).

Both of those scenarios are probably not what you'd expect to happen.

To fix this, this removes the use of pkg-config when --with-openssl-dir
is specified.

To further make sure that --with-openssl-dir functions as the override it
was meant to be, the check for a "lib" directory under the specified
location is changed to a more comprehensive check for the library itself.

It uses 'find' utility to look for "*crypto.*", which should find it for
most naming schemes (it will find "libcrypto.a", "libcrypto.so",
"crypto.dll", etc.).

The 'find' syntax used should be POSIX compliant. It has been tested with
GNU, BSD and Solaris versions of 'find'.

Since OpenSSL dependencies were handled by pkg-config, those have to now
be added unconditionally when --with-openssl-dir is specified (in addition
to when pkg-config isn't available).

This may result in a extraneous zlib (if OpenSSL was built without it)
appearing on the link line (ref. "pedantic reasons" above). Modern linkers
will not record a dependency on this unless explicitly told to do so.