configure.ac: Drop pkg-config usage when --with-nss-dir is specified
When installing NSS (using 'make install'), it's possible to specify a
DESTDIR which will install it to that location instead of the location
specified with --prefix during configuration.

However, package config files are generated based on the --prefix
location, so even if installed to somewhere else, the package config file
will still point to the final destination provided during configuration.

This configure script would call pkg-config, if available, also
when --with-nss-dir was used. This is primarily because the --with-nss-dir
handling is a copy-and-paste of the --with-openssl-dir handling, which
also used to do this. That was changed in PR# 481. This is essentially
doing the same thing for the --with-nss-dir flags.

If DESTDIR was used to install NSS, pkg-config would add flags
pointing to the --prefix location used when building NSS, not the
provided --with-nss-dir.

This could end up picking up a system library instead, or, if there
happened to be a NSS library already at the given --prefix, that would
be used.

Both of those scenarios are probably not what you'd expect to happen.

To fix this, this removes the use of pkg-config when --with-nss-dir
is specified.

To further make sure that --with-nss-dir functions as the override it
was meant to be, the check for a "lib" directory under the specified
location is changed to a more comprehensive check for the library itself.

It uses 'find' utility to look for "*nss3.*", which should find it for
most naming schemes (it will find "libnss3.a", "libnns3.so", "nss3.dll",
etc.).

The 'find' syntax used should be POSIX compliant. It has been tested
with GNU, BSD and Solaris versions of 'find'.

Note that this could break existing build configurations that uses
--with-nss-dir to point to non-system package intall location.

For instance, MacPorts installs NSS and NSPR to /opt/local, but in
subdirectories. So include files go in /opt/local/include/[nss|nspr] and
libraries go in /opt/local/lib/[nss|nspr].

If you tried to use --with-nss-dir=/opt/local, this would fail to find a
NSS library in /opt/local/lib, whereas the old method would find the
nss.pc file in /opt/local/lib/pkgconfig and pick up the correct flags from
running pkg-config.

If you're in this situation, with this change, you would have to set
PKG_CONFIG_PATH before running configure instead of --with-nss-dir.