test/aes_calc: Support checking the ciphertext
Modify aes_calc to optionally accept the expected ciphertext as
a third argument, so it can error with a non-0 exit code if the
calculated ciphertext doesn't match the expected ciphertext. This
makes our test setup with meson much easier, as we don't need to parse
and compare the output of the test binary.

The test is now self-sufficient.