configure: check OPENSSL_cleanse() behavior based OpenSSL version
As reported in #479, testing for OPENSSL_cleanse()
behavior within ./configure cannot happen when used
in cross compilation environment.

The initial issue addressed by this runtime test in
./configure was reported in #414 with OPENSSL_cleanse()
and was said to be related to OpenSSL 1.0.2g on aarch64.

Subsequent releases of OpenSSL address the issue, and
should be considered fixed as of:
- OpenSSL 1.0.2i,
  with commit 5bbdc26cadc01cab811040e861f1f98e0f3af348 ("crypto/mem_clr.c: switch to OPENSSL_cleanse implementation from master.")
- OpenSSL 1.1.0 and up,
  with commit 104ce8a9f02d250dd43c255eb7b8747e81b29422 ("RT4116: Change cleanse to just memset")

Then there's no reason for current OpenSSL versions to
use the broken OPENSSL_cleanse() implementation, so the
runtime test is almost useless and can be replaced by
a version check.

If older OpenSSL version is detected, runtime OPENSSL_cleanse()
test will take place as before (provided libsrtp is not
to be cross compiled). If newer OpenSSL version is detected,
no runtime OPENSSL_cleanse() is needed.