suport running against nss >= 3.52
Fix for #523

This is described in more detail at:
https://bugzilla.mozilla.org/show_bug.cgi?id=1637488#c3

Have decided to go with using NSS_PKCS11_2_0_COMPAT for now
as it provides the most fexibility.
Consider moving to the new PK11_AEADOp api in the future but
that will put a requirment on nss version at both compile and
run time.