alternative stream list api and unit tests for the api
Beside some small API rename the main difference is that the
list implementation is not responsible for deallocating streams.
Calling srtp_stream_list_dealloc() on a list that is not empty
results in an error.

This also changes the default implementation to be a double linked
list making removing items faster.