iv length is constant so set only once
The iv length is preserved inside the EVP_CIPHER_CTX
so no need to set more than once.
This is especially important with OpenSSL 3 where setting the iv len is a expensive
operation due to param lookup code inside of OpenSSL.

This should help Performance issue with OpennSSL 3 and libsrtp with #645 in the case of GCM
but does not fix all of the performance issues.