srtp_policy: add force_zero_roc policy flag
Note that if the ROC is forced to zero, only 2^16 packets
can be sent with a given (master or session) key or a severe
security weakness is introduced!

See Section 3.3.1 of RFC 3711:
Each time the RTP sequence number, SEQ, wraps
   modulo 2^16, the sender side MUST increment ROC by one, modulo 2^32.

This patch is added to cover an old and wrong srtp implementation where
no authentication (and related to that), no ROC increment was done.

So, please don't use it, unless you are for legacy reasons enforced
to do so.

Signed-off-by: Thomas Devoogdt <thomas.devoogdt@barco.com>