enable cryptex support
Move cryptex code to separate functions so it can be reused.
Due to support for non in-place io it is not always possible to modify the
input buffer in the way the cryptex draft expected therefore use multiple
encrypt / decrypt calls when io is not in-place. The gcm ciphers do not
currently support multiple operations so non in-place io with csrc's is not
supported.