add cipher test that verifies multiple packets
libSRTP will allocate and initialize a cipher once and then use it
to encrypt or decrypt multiple packets. Between each operation
a new IV will be set, the IV is unique for each packet. This test will
ensure that cipher implementations support this.

The test vectors are based on examples in RFC 9335 .