Properly support null crypto and null auth scenario
Currently, it is silently assumed that we always have master key provided,
even when the policy says we don't use encryption and authentication. This
results in streams with such policy to fail.

Fix that by properly calculating the expected master key length by taking
in account authentication type too. Correctly set expected key length for
null crypto policy to 0. Do not assume we have HMAC if the policy is set to
not have.

Add a test case for null/null policy while at it.

Signed-off-by: Ivaylo Dimitrov <ivo.g.dimitrov.75@gmail.com>